cache enabled

+++ src/main/java/yandex/eventservice/domain/Event.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 public class Event {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Long id;
 
     private LocalDateTime date;
 
@@ -19,11 +19,11 @@ public class Event {
         this.date = date;
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
